backend/pyproject.toml

[tool.poetry]
name = "gratitude-network-backend"
version = "0.1.0"
description = "Backend for Gratitude Network"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
fastapi = "==0.110.0"
uvicorn = "==0.22.0"
python-dotenv = "==0.21.0"
psycopg2-binary = "==2.9.10"
sqlmodel = "==0.0.16"
Pillow = "==10.3.0"
python-multipart = "==0.0.9"
bcrypt = "==4.3.0"
python-jose = "==3.5.0"
email-validator = "*"
SQLAlchemy = "==2.0.41"
    passlib = {extras = ["bcrypt"], version = "*"}
redis = "*"

[tool.poetry.group.dev.dependencies]
pytest = "==7.2.0"
httpx = "==0.27.0"
pytest-asyncio = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]

